# modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/rest-openapi.yaml
components:
    schemas:
        Assignee:
            description:
                https://schema.org/Assignee
            properties:
                id:
                    description:
                        The user's ID.
                    format: int64
                    readOnly: true
                    type: integer
                image:
                    description:
                        A relative URL to the user's profile image.
                    format: uri
                    readOnly: true
                    type: string
                name:
                    description:
                        The user's full name.
                    readOnly: true
                    type: string
            type: object
...
info:
    description:
        ""
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: ""
    version: v1.0
openapi: 3.0.1
paths:
...
    "/processes/{processId}/assignees":
        get:
            operationId: getProcessAssigneesPage
            parameters:
                - in: path
                  name: processId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Assignee"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Assignee"
                                type: array
                    description:
                        ""
            tags: ["Assignee"]
...